+++ age/src/protocol.rs
@@ -214,10 +214,15 @@ impl<R: AsyncRead + Unpin> Decryptor<R> {
 mod tests {
     use secrecy::SecretString;
     use std::io::{BufReader, Read, Write};
+
+    #[cfg(feature = "ssh")]
     use std::iter;
 
     use super::{Decryptor, Encryptor};
-    use crate::{identity::IdentityFile, x25519, Identity, Recipient};
+    use crate::{
+        identity::{IdentityFile, IdentityFileEntry},
+        x25519, Identity, Recipient,
+    };
 
     #[cfg(feature = "async")]
     use futures::{
@@ -339,7 +344,11 @@ mod tests {
         let pk: x25519::Recipient = crate::x25519::tests::TEST_PK.parse().unwrap();
         recipient_round_trip(
             vec![Box::new(pk)],
-            f.into_identities().iter().map(|sk| sk as &dyn Identity),
+            f.into_identities().iter().map(|sk| match sk {
+                IdentityFileEntry::Native(sk) => sk as &dyn Identity,
+                #[cfg(feature = "plugin")]
+                IdentityFileEntry::Plugin(_) => unreachable!(),
+            }),
         );
     }
 
@@ -351,7 +360,11 @@ mod tests {
         let pk: x25519::Recipient = crate::x25519::tests::TEST_PK.parse().unwrap();
         recipient_async_round_trip(
             vec![Box::new(pk)],
-            f.into_identities().iter().map(|sk| sk as &dyn Identity),
+            f.into_identities().iter().map(|sk| match sk {
+                IdentityFileEntry::Native(sk) => sk as &dyn Identity,
+                #[cfg(feature = "plugin")]
+                IdentityFileEntry::Plugin(_) => unreachable!(),
+            }),
         );
     }